Remove keyword param from request
This may seem like an anti-pattern for most APIs that require the data
for to go under a key of the same name in the body of a response.

We're assuming since that name exists in the URI we don't need it twice.

For example:

curl -X POST -H "Content-Type: application/json" \
     -d '{"title": "foo", "color": "red", "priority": "high"}' \
     http://localhost:3000/things

This would create a new thing with those attributes.

curl -X PUT -H "Content-Type: application/json" \
     -d '{"title": "fixed", "color": "yellow", "priority": "medium"}' \
     http://localhost:3000/things/2

Would update the existing thing with ID#2 given the provided map.

Without this change, you would have to do something like this:

curl -X POST -H "Content-Type: application/json" \
     -d '{"thing":
	   {"title": "bar", "color": "green", "priority": "low"}
	 }' \
     http://localhost:3000/things

I'm not sure which I prefer but certainly less typing is involved with
this patch, but maybe there is an argument for expecting other stuff
in the body?